web-animations/interfaces/KeyframeEffect/mutable-effects.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Tests for modifying animation effects</title>
<link rel="help" href="https://w3c.github.io/web-animations/#the-keyframeeffect-interfaces">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
'use strict';

test(function() {
  document.timeline.playbackRate = 0;
  document.timeline.currentTime = 0;
  var player = document.documentElement.animate([], 1000);
  player.startTime = -1000;
  assert_equals(player.currentTime, 1000);
  player.effect.timing.duration = 1500;
  document.timeline.currentTime = 2000;
  assert_equals(player.startTime, -1000);
  assert_equals(player.currentTime, 1500);
}, 'Extending effect should not cause a change in start time.');

</script>
</body>
